+++ cli/tests/pcluster/schemas/test_common_schema.py
@@ -10,8 +10,9 @@
 # limitations under the License.
 import pytest
 from assertpy import assert_that
+from marshmallow import ValidationError
 
-from pcluster.schemas.common_schema import validate_json_format
+from pcluster.schemas.common_schema import ImdsSchema, validate_json_format
 
 
 @pytest.mark.parametrize(
@@ -23,3 +24,25 @@ from pcluster.schemas.common_schema import validate_json_format
 )
 def test_validate_json_format(data, expected_value):
     assert_that(validate_json_format(data)).is_equal_to(expected_value)
+
+
+@pytest.mark.parametrize(
+    "imds_support, failure_message",
+    [
+        ("unsupportedValue", "Must be one of"),
+        ("v1.0", None),
+        ("v2.0", None),
+    ],
+)
+def test_imds_schema(imds_support, failure_message):
+    imds_schema = {}
+
+    if imds_support:
+        imds_schema["ImdsSupport"] = imds_support
+
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ImdsSchema().load(imds_schema)
+    else:
+        imds = ImdsSchema().load(imds_schema)
+        assert_that(imds.imds_support).is_equal_to(imds_support)